+++ HISTORY.md
@@ -11,6 +11,8 @@
 * Add `Wavefront::Account#admins`.
 * Add `Wavefront::Account#update_perms`.
 * Add `Wavefront::Integration#manifest_min`.
+* Add `Wavefront::Proxy#config`.
+* Add `Wavefront::Proxy#preprocessor_rules`.
 
 ## 7.0.0 (2023-01-19)
 * Drop support for Ruby 2.5. (Breaking change.)